+++ mex/common/sinks/backend_api.py
@@ -1,11 +1,9 @@
 from collections.abc import Generator, Iterable
-from typing import cast
 
 from mex.common.backend_api.connector import BackendApiConnector
 from mex.common.logging import logger
-from mex.common.models import AnyExtractedModel
+from mex.common.models import AnyExtractedModel, AnyMergedModel, AnyRuleSetResponse
 from mex.common.sinks.base import BaseSink
-from mex.common.types import AnyExtractedIdentifier
 from mex.common.utils import grouper
 
 
@@ -15,22 +13,30 @@ class BackendApiSink(BaseSink):
     CHUNK_SIZE = 50
 
     def load(
-        self,
-        models: Iterable[AnyExtractedModel],
-    ) -> Generator[AnyExtractedIdentifier, None, None]:
-        """Load models to the Backend API using bulk insertion.
+        self, items: Iterable[AnyExtractedModel | AnyMergedModel | AnyRuleSetResponse]
+    ) -> Generator[AnyExtractedModel | AnyMergedModel | AnyRuleSetResponse, None, None]:
+        """Load extracted models or rule-sets to the Backend API using bulk insertion.
 
         Args:
-            models: Iterable of extracted models
+            items: Iterable of extracted models or merged models or rule-sets
+
+        Raises:
+            NotImplementedError: When you try to load merged items into the backend
 
         Returns:
-            Generator for identifiers of posted models
+            Generator for posted models
         """
         total_count = 0
         connector = BackendApiConnector.get()
-        for chunk in grouper(self.CHUNK_SIZE, models):
-            model_list = [model for model in chunk if model is not None]
-            response = connector.post_extracted_items(model_list)
+        for chunk in grouper(self.CHUNK_SIZE, items):
+            model_list = []
+            for model in chunk:
+                if isinstance(model, AnyExtractedModel | AnyRuleSetResponse):
+                    model_list.append(model)
+                elif model is not None:
+                    msg = f"backend cannot ingest {type(model)}"
+                    raise NotImplementedError(msg)
+            connector.ingest(model_list)
             total_count += len(model_list)
-            yield from cast(list[AnyExtractedIdentifier], response.identifiers)
+            yield from model_list
             logger.info("%s - written %s models", type(self).__name__, total_count)