+++ mex/common/sinks/ndjson.py
@@ -2,15 +2,16 @@ import json
 from collections.abc import Generator, Iterable
 from contextlib import ExitStack
 from pathlib import Path
-from typing import IO, Any
+from typing import IO, Any, TypeVar
 
 from mex.common.logging import logger
-from mex.common.models import AnyExtractedModel
+from mex.common.models.base.model import BaseModel
 from mex.common.settings import BaseSettings
 from mex.common.sinks.base import BaseSink
 from mex.common.transform import MExEncoder
-from mex.common.types import AnyExtractedIdentifier
 from mex.common.utils import grouper
+
+_LoadItemT = TypeVar("_LoadItemT", bound=BaseModel)
 
 
 class NdjsonSink(BaseSink):
@@ -29,24 +30,24 @@ class NdjsonSink(BaseSink):
 
     def load(
         self,
-        models: Iterable[AnyExtractedModel],
-    ) -> Generator[AnyExtractedIdentifier, None, None]:
-        """Write models into a new-line delimited JSON file.
+        items: Iterable[_LoadItemT],
+    ) -> Generator[_LoadItemT, None, None]:
+        """Write any items into a new-line delimited JSON file.
 
         Args:
-            models: Iterable of extracted models to write
+            items: Iterable of any kind of items
 
         Returns:
-            Generator for identifiers of written models
+            Generator for the loaded items
         """
         file_handles: dict[str, IO[Any]] = {}
         total_count = 0
         with ExitStack() as stack:
-            for chunk in grouper(self.CHUNK_SIZE, models):
-                for model in chunk:
-                    if model is None:
+            for chunk in grouper(self.CHUNK_SIZE, items):
+                for item in chunk:
+                    if item is None:
                         continue
-                    class_name = model.__class__.__name__
+                    class_name = item.__class__.__name__
                     try:
                         fh = file_handles[class_name]
                     except KeyError:
@@ -59,7 +60,8 @@ class NdjsonSink(BaseSink):
                             class_name,
                             file_name.as_posix(),
                         )
-                    fh.write(f"{json.dumps(model, sort_keys=True, cls=MExEncoder)}\n")
+                    dumped_json = json.dumps(item, sort_keys=True, cls=MExEncoder)
+                    fh.write(f"{dumped_json}\n")
                     total_count += 1
-                    yield model.identifier
-                logger.info("%s - written %s models", type(self).__name__, total_count)
+                    yield item
+                logger.info("%s - written %s items", type(self).__name__, total_count)