+++ tests/sinks/test_backend_api.py
@@ -1,10 +1,10 @@
 from unittest.mock import MagicMock, Mock
 
+import pytest
 from pytest import MonkeyPatch
 
 from mex.common.backend_api.connector import BackendApiConnector
-from mex.common.backend_api.models import IdentifiersResponse
-from mex.common.models import ExtractedPerson
+from mex.common.models import ExtractedPerson, ItemsContainer, MergedPerson
 from mex.common.sinks.backend_api import BackendApiSink
 
 
@@ -16,13 +16,24 @@ def test_sink_load_mocked(
 
     monkeypatch.setattr(BackendApiConnector, "__init__", __init__)
 
-    response = IdentifiersResponse(identifiers=[extracted_person.identifier])
-    post_extracted_items = Mock(return_value=response)
-    monkeypatch.setattr(
-        BackendApiConnector, "post_extracted_items", post_extracted_items
-    )
+    response = ItemsContainer[ExtractedPerson](items=[extracted_person])
+    ingest = Mock(return_value=response)
+    monkeypatch.setattr(BackendApiConnector, "ingest", ingest)
 
     sink = BackendApiSink.get()
-    model_ids = list(sink.load([extracted_person]))
-    assert model_ids == response.identifiers
-    post_extracted_items.assert_called_once_with([extracted_person])
+    models_or_rule_sets = list(sink.load([extracted_person]))
+    assert models_or_rule_sets == [extracted_person]
+    ingest.assert_called_once_with([extracted_person])
+
+
+def test_sink_load_merged_error(
+    merged_person: MergedPerson, monkeypatch: MonkeyPatch
+) -> None:
+    def __init__(self: BackendApiConnector) -> None:
+        self.session = MagicMock()
+
+    monkeypatch.setattr(BackendApiConnector, "__init__", __init__)
+
+    sink = BackendApiSink.get()
+    with pytest.raises(NotImplementedError, match="backend cannot"):
+        list(sink.load([merged_person]))